:sparkles: feature에 채팅방, 채팅 분리
