+++ src/features/chat-room/api/websocket/WebSocketServiceImpl.ts
@@ -1,7 +1,7 @@
 import { Client } from "@stomp/stompjs";
 import SockJS from "sockjs-client";
 import { WebSocketService, WebSocketMessage, TypingIndicatorMessage, MessageStatusUpdate } from "./types";
-import { ChatMessageItem } from "../../model/types/ChatRoom.types";
+import { ChatMessageItem } from "../../../message/model/types/ChatRoom.types";
 
 export class WebSocketServiceImpl implements WebSocketService {
     private client: Client | null = null;
