--- src/features/chat/api/index.ts
@@ -1,5 +1,0 @@
-export * from './chatRoom';
-export * from './message';
-export * from './messageReaction';
-export * from './websocket';
-export * from './websocket/types';
