+++ src/features/message/model/types/ChatRoom.types.ts
