+++ src/features/social/ui/FriendsTab.tsx
@@ -3,7 +3,7 @@ import { useAuth } from "../../../shared/lib/context/AuthContext";
 import { useFriends } from "../model/hooks/useFriends";
 import FriendSearch from "../../user-code/ui/friendSearch";
 import FriendCodePage from "../../user-code/ui/friendCodePage";
-import { createDirectChat } from "../../chat/api/chatRoom";
+import { createDirectChat } from "../../chat-room/api/chatRoom";
 import { useNavigate } from "react-router-dom";
 import TabContainer from "../../../shared/ui/TabContainer";
 import TabHeader from "../../../shared/ui/TabHeader";
