+++ src/pages/chat/ChatRoomPage.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { useParams } from 'react-router-dom';
-import ChatRoom from '../../features/chat/ChatRoom';
+import ChatRoom from '../../features/chat-room/ChatRoom';
 
 const ChatRoomPage: React.FC = () => {
-  const { roomId } = useParams<{ roomId: string }>();
-  if (!roomId) return null;
-  return <ChatRoom roomId={roomId} />;
+    const { roomId } = useParams<{ roomId: string }>();
+    if (!roomId) return null;
+    return <ChatRoom roomId={roomId} />;
 };
 
 export default ChatRoomPage;
